Update split routine
Consider the state of the B+ Tree of degree 2 be a leaf node containing
keys [1,2,3]. On inserting 4, the root node is modified to contain [2]
and the left and right child contain [1,2] and [3,4] respectively. When
evaluating the leaf index of say 3, in the first case it is 0 but in the
second case it is 1. This is solved by splitting the node at degree'th
index instead of (degree-1).

Evaluation of leaf indices will be needed to find the corresponding disk
blocks. It should be noted that the above argument holds only for sequential and
insertions ; random insertions are not yet supported.